various: ımprove attribute path validation & error handling across platforms
Makes `toplevel_for` return `Result` types for all platforms to properly
handle and report invalid attribute paths. This also means that we get
to infer correct installables from previously "invalid" installables
such as `flake.nix#myHost` where this is normalized to
`path/to/flake.nix` and the hostname `myHost` is automatically passed on
to the toplevel handler.

Additionally:

- Fix Darwin validation to catch all invalid paths, not just those with "config" at position 2
- Add re-validation in Home Manager after prepending homeConfigurations to catch invalid paths
- Move Darwin's toplevel_for from NixOS import to platform-specific implementation
- Remove unnecessary attribute preprocessing in Darwin rebuild and update call site

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: Icb31f494e4734611d704ffbdea300c256a6a6964