commands: fix error handling in nom pipeline execution
Fixes a minor issue in how commands that are invalid or improperly
handled are forwarded to the Nix command. Replaces `join()` with
`popen()` to access individual processes in the pipeline. This way we
can better check the exist status of the `nix build` process and
properly propagate them.

Also improves naming a little bit because why not?

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: I8a44abf924f9c9a1c06d102e5a3f40aa6a6a6964