remote: properly quote SSH command arguments; improve error handling
I've cramped too many things in one commit again. Guess it's better to
explain, generally, what has changed.

The primary "fix" that this commit addresses is SSH command construction
by shell-quoting *all* arguments before passing them to the remote
shell, instead of using SSH's `'--'` separator with individual args.
This helps ensure special chars, spaces and shell metachars in paths (I
blame Nix) or arguments args are handled correctly.

Additionally, there are a few improvements to the "robustness" of remote
copying: previously we'd attempt to copy from build-host to target-host,
which is fine, but we failed fast instead of attempting the logically
correct behaviour which is to COPY BACK and THEN relay to the target
host. If remote-to-remote copy fails, the system now logs a warning and
relays through localhost instead of failing. This makes nh's *fallback
behaviour* (but not the first pass) consistent with nixos-rebuild.

Other than that, some build command consolidation and moderately large
documentation tweaks. The code is getting out of hand, so I'd like to
merge this and be done as soon as possible.

God that was a long summary. Hope someone gets to read this.

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: Ib540fc287f1b1e22e2b68b9c9c7b03046a6a6964